+++ aten/src/TH/CMakeLists.txt
@@ -36,6 +36,7 @@ set(ATen_TH_SRCS
   ${CMAKE_CURRENT_SOURCE_DIR}/THTensorRandom.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/THTensorMath.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/THTensorMoreMath.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/THTensorEvenMoreMath.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/THTensorConv.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/THTensorLapack.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/THBlas.cpp
