+++ aten/src/TH/THTensorEvenMoreMath.cpp
@@ -1,0 +1,7 @@
+#include "THTensor.hpp"
+#include "THVector.h"
+#include "THBlas.h"
+#include "THTensorDimApply.h"
+
+#include "generic/THTensorEvenMoreMath.cpp"
+#include "THGenerateAllTypes.h"
